Link bg_clip.c

+++ include/connection.h
@@ -10,7 +10,7 @@ void ConnectionUpdateHatchAnimation(u8 dontSetRaw, u32 hatch);
 void ConnectionHatchFlashingAnimation(u8 hatch);
 void ConnectionOverrideOpenedHatch(u8 hatch, u32 type);
 u32 ConnectionCheckEnterDoor(u16 yPosition, u16 xPosition);
-u8 ConnectionCheckAreaConnection(u16 yPosition, u16 xPosition);
+u32 ConnectionCheckAreaConnection(u16 yPosition, u16 xPosition);
 void ConnectionProcessDoorType(u8 type);
 u8 ConnectionFindEventBasedDoor(u8 sourceDoor);
 u32 ConnectionSetHatchAsOpened(u8 action, u8 hatch);
